// src/generated/linear-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  DateTimeOrDuration: { input: any; output: any; }
  Duration: { input: any; output: any; }
  JSON: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  TimelessDate: { input: any; output: any; }
  TimelessDateOrDuration: { input: any; output: any; }
  UUID: { input: any; output: any; }
};

/** A bot actor is an actor that is not a user, but an application or integration. */
export type ActorBot = {
  __typename?: 'ActorBot';
  /** A url pointing to the avatar representing this bot. */
  avatarUrl?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  /** The display name of the bot. */
  name?: Maybe<Scalars['String']['output']>;
  /** The sub type of the bot. */
  subType?: Maybe<Scalars['String']['output']>;
  /** The type of bot. */
  type: Scalars['String']['output'];
  /** The display name of the external user on behalf of which the bot acted. */
  userDisplayName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Comment = Node & {
  __typename?: 'Comment';
  /** The time at which the entity was archived. Null if the entity has not been archived. */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The comment content in markdown format. */
  body: Scalars['String']['output'];
  /** [Internal] The comment content as a Prosemirror document. */
  bodyData: Scalars['String']['output'];
  /** The bot that created the comment. */
  botActor?: Maybe<ActorBot>;
  /** The children of the comment. */
  children: CommentConnection;
  /** The time at which the entity was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The document content that the comment is associated with. */
  documentContent?: Maybe<DocumentContent>;
  /** The time user edited the comment. */
  editedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The external thread that the comment is synced with. */
  externalThread?: Maybe<SyncedExternalThread>;
  /** The external user who wrote the comment. */
  externalUser?: Maybe<ExternalUser>;
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
  /** The initiative update that the comment is associated with. */
  initiativeUpdate?: Maybe<InitiativeUpdate>;
  /** The issue that the comment is associated with. */
  issue?: Maybe<Issue>;
  /** The parent comment under which the current comment is nested. */
  parent?: Maybe<Comment>;
  /** The post that the comment is associated with. */
  post?: Maybe<Post>;
  /** The project update that the comment is associated with. */
  projectUpdate?: Maybe<ProjectUpdate>;
  /** The text that this comment references. Only defined for inline comments. */
  quotedText?: Maybe<Scalars['String']['output']>;
  /** Emoji reaction summary, grouped by emoji type. */
  reactionData: Scalars['JSONObject']['output'];
  /** Reactions associated with the comment. */
  reactions: Array<Reaction>;
  /** The time the resolvingUser resolved the thread. */
  resolvedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The comment that resolved the thread. */
  resolvingComment?: Maybe<Comment>;
  /** The user that resolved the thread. */
  resolvingUser?: Maybe<User>;
  /** [Internal] A generated summary of the comment thread. */
  threadSummary?: Maybe<Scalars['JSONObject']['output']>;
  /**
   * The last time at which the entity was meaningfully updated. This is the same as the creation time if the entity hasn't
   *     been updated after creation.
   */
  updatedAt: Scalars['DateTime']['output'];
  /** Comment's URL. */
  url: Scalars['String']['output'];
  /** The user who wrote the comment. */
  user?: Maybe<User>;
};
// ...
export type CommentConnection = {
  __typename?: 'CommentConnection';
  edges: Array<CommentEdge>;
  nodes: Array<Comment>;
  pageInfo: PageInfo;
};
// ...
export type CommentEdge = {
  __typename?: 'CommentEdge';
  /** Used in `before` and `after` args */
  cursor: Scalars['String']['output'];
  node: Comment;
};
// ...
export type DocumentContent = Node & {
  __typename?: 'DocumentContent';
  /** The time at which the entity was archived. Null if the entity has not been archived. */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The document content in markdown format. */
  content?: Maybe<Scalars['String']['output']>;
  /** The document content state as a base64 encoded string. */
  contentState?: Maybe<Scalars['String']['output']>;
  /** The time at which the entity was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The document that the content is associated with. */
  document?: Maybe<Document>;
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
  /** The initiative that the content is associated with. */
  initiative?: Maybe<Initiative>;
  /** The issue that the content is associated with. */
  issue?: Maybe<Issue>;
  /** [ALPHA] The meeting that the content is associated with. */
  meeting?: Maybe<Meeting>;
  /** The project that the content is associated with. */
  project?: Maybe<Project>;
  /** The project milestone that the content is associated with. */
  projectMilestone?: Maybe<ProjectMilestone>;
  /** The time at which the document content was restored from a previous version. */
  restoredAt?: Maybe<Scalars['DateTime']['output']>;
  /**
   * The last time at which the entity was meaningfully updated. This is the same as the creation time if the entity hasn't
   *     been updated after creation.
   */
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type ExternalUser = Node & {
  __typename?: 'ExternalUser';
  /** The time at which the entity was archived. Null if the entity has not been archived. */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** An URL to the external user's avatar image. */
  avatarUrl?: Maybe<Scalars['String']['output']>;
  /** The time at which the entity was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The external user's display name. Unique within each organization. Can match the display name of an actual user. */
  displayName: Scalars['String']['output'];
  /** The external user's email address. */
  email?: Maybe<Scalars['String']['output']>;
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
  /** The last time the external user was seen interacting with Linear. */
  lastSeen?: Maybe<Scalars['DateTime']['output']>;
  /** The external user's full name. */
  name: Scalars['String']['output'];
  /** Organization the external user belongs to. */
  organization: Organization;
  /**
   * The last time at which the entity was meaningfully updated. This is the same as the creation time if the entity hasn't
   *     been updated after creation.
   */
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type InitiativeUpdate = Node & {
  __typename?: 'InitiativeUpdate';
  /** The time at which the entity was archived. Null if the entity has not been archived. */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The update content in markdown format. */
  body: Scalars['String']['output'];
  /** [Internal] The content of the update as a Prosemirror document. */
  bodyData: Scalars['String']['output'];
  /** Comments associated with the initiative update. */
  comments: CommentConnection;
  /** The time at which the entity was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The diff between the current update and the previous one. */
  diff?: Maybe<Scalars['JSONObject']['output']>;
  /** The diff between the current update and the previous one, formatted as markdown. */
  diffMarkdown?: Maybe<Scalars['String']['output']>;
  /** The time the update was edited. */
  editedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The health at the time of the update. */
  health: InitiativeUpdateHealthType;
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
  /** [Internal] Serialized JSON representing current state of the initiative properties when posting the initiative update. */
  infoSnapshot?: Maybe<Scalars['JSONObject']['output']>;
  /** The initiative that the update is associated with. */
  initiative: Initiative;
  /** Whether initiative update diff should be hidden. */
  isDiffHidden: Scalars['Boolean']['output'];
  /** Whether the initiative update is stale. */
  isStale: Scalars['Boolean']['output'];
  /** Emoji reaction summary, grouped by emoji type. */
  reactionData: Scalars['JSONObject']['output'];
  /** Reactions associated with the initiative update. */
  reactions: Array<Reaction>;
  /** The update's unique URL slug. */
  slugId: Scalars['String']['output'];
  /**
   * The last time at which the entity was meaningfully updated. This is the same as the creation time if the entity hasn't
   *     been updated after creation.
   */
  updatedAt: Scalars['DateTime']['output'];
  /** The URL to the initiative update. */
  url: Scalars['String']['output'];
  /** The user who wrote the update. */
  user: User;
};
// ...
export enum InitiativeUpdateHealthType {
  AtRisk = 'atRisk',
  OffTrack = 'offTrack',
  OnTrack = 'onTrack'
}
// ...
export type Node = {
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
};
// ...
export type Reaction = Node & {
  __typename?: 'Reaction';
  /** The time at which the entity was archived. Null if the entity has not been archived. */
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  /** The comment that the reaction is associated with. */
  comment?: Maybe<Comment>;
  /** The time at which the entity was created. */
  createdAt: Scalars['DateTime']['output'];
  /** Name of the reaction's emoji. */
  emoji: Scalars['String']['output'];
  /** The external user that created the reaction. */
  externalUser?: Maybe<ExternalUser>;
  /** The unique identifier of the entity. */
  id: Scalars['ID']['output'];
  /** The initiative update that the reaction is associated with. */
  initiativeUpdate?: Maybe<InitiativeUpdate>;
  /** The issue that the reaction is associated with. */
  issue?: Maybe<Issue>;
  /** The post that the reaction is associated with. */
  post?: Maybe<Post>;
  /** The project update that the reaction is associated with. */
  projectUpdate?: Maybe<ProjectUpdate>;
  /**
   * The last time at which the entity was meaningfully updated. This is the same as the creation time if the entity hasn't
   *     been updated after creation.
   */
  updatedAt: Scalars['DateTime']['output'];
  /** The user that created the reaction. */
  user?: Maybe<User>;
};
// ...
export type SyncedExternalThread = {
  __typename?: 'SyncedExternalThread';
  /** The display name of the thread. */
  displayName?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  /** Whether this thread is syncing with the external service. */
  isConnected: Scalars['Boolean']['output'];
  /** Whether the current user has the corresponding personal integration connected for the external service. */
  isPersonalIntegrationConnected: Scalars['Boolean']['output'];
  /** Whether a connected personal integration is required to comment in this thread. */
  isPersonalIntegrationRequired: Scalars['Boolean']['output'];
  /** The display name of the source. */
  name?: Maybe<Scalars['String']['output']>;
  /** The sub type of the external source. */
  subType?: Maybe<Scalars['String']['output']>;
  /** The type of the external source. */
  type: Scalars['String']['output'];
  /** The external url of the thread. */
  url?: Maybe<Scalars['String']['output']>;
};